LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-calcMode-discrete-path-curves-expected.txt


FAIL Test animateMotion with calcMode="discrete" and complex paths assert_equals: curves: moveTo (x) expected 0 but got 33.828853607177734
